refactor(#442): split DigitalTeachingToolbar into components + shared hooks
Behaviour-preserving refactor. The single 1679-line
components/teachingTools/DigitalTeachingToolbar.tsx is split into:
- DigitalTeachingToolbar.tsx (398 lines) — toolbar shell only
- TimerTool.tsx / DiceTool.tsx / RpsTool.tsx / DieFace.tsx
- hooks/useDraggable.ts, hooks/useResizable.ts (+ getMaxScale)
- types.ts (shared ToolProps)

The ~310 lines of duplicated startDrag/startResize are unified into two
hooks. The three copies were NOT identical, so the hooks are parameterized
to preserve each tool's exact behaviour rather than flattening it:
- useDraggable(extraIgnoreSelectors): always ignores button + .resize-handle;
  Timer adds [".settings-panel"], Dice adds [".dice-clickable"], RPS none.
- useResizable(containerRef, fallbackW, fallbackH, minScale): Timer
  (containerRef,280,320,0.5); Dice (diceContainerRef,200,200,0.8); RPS
  (rpsContainerRef,220,260,0.8). Timer's min scale is 0.5, Dice/RPS 0.8 —
  preserved via the minScale param. Dice originally hardcoded
  getMaxScale(200,200); the hook reads diceContainerRef.offsetWidth ?? 200
  on its fixed 200x200 container, which yields the same 200 (offsetWidth is
  the untransformed layout size).

The default export and the TimerTool/DiceTool/RpsTool props
(show/onClose/zCounterRef) are unchanged, so importers and the existing
test file are unaffected.

Verification (code-equivalence + build): the two hooks are byte-identical to
the original startDrag/startResize bodies; call-site params verified against
the originals; tsc, eslint (--max-warnings 0), prettier, and vite build all
pass; the RpsTool regression tests (#443) still pass and the toolbar test
file result is unchanged from before the refactor (the pre-existing
userEvent Timer/Dice local failures are unrelated and pass in CI). The
vitest suite mounts the full component and opens all three tools without
runtime error. A live drag/resize gesture test needs the authed teacher app
(backend + login) and is best eyeballed on the staging preview.

Closes #442
