+++ frontend/src/components/teachingTools/types.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export interface ToolProps {
+  show: boolean;
+  onClose: () => void;
+  zCounterRef: React.MutableRefObject<number>;
+}
